Docs and fixes
